+++ jaguar/lib/jaguar.dart
@@ -6,3 +6,4 @@ export 'serve/server.dart';
 export 'http/http.dart';
 export 'utils/string/import.dart';
 export 'package:path_tree/path_tree.dart' show pathToSegments, cleanupSegments;
+export 'dart:io' show SecurityContext;